+++ IOPool/Streamer/interface/StreamerOutSrvcManager.h
@@ -1,7 +1,7 @@
 #ifndef _STREAMEROUTSRVCMANAGER_H_
 #define _STREAMEROUTSRVCMANAGER_H_
 
-// $Id:$
+// $Id: StreamerOutSrvcManager.h,v 1.8 2006/11/29 10:05:00 klute Exp $
 
 #include "FWCore/ParameterSet/interface/ProcessDesc.h"
 #include "FWCore/Framework/interface/EventSelector.h"
@@ -32,7 +32,7 @@ namespace edm
     
     void stop(); 
     
-    void manageInitMsg(std::string fileName, uint32 runNum, unsigned long maxFileSize, double highWaterMark,
+    void manageInitMsg(std::string fileName, uint32 runNum, unsigned int maxFileSize, double highWaterMark,
 		       std::string path, std::string mpath, std::string catalog, uint32 disks, 
 		       InitMsgView& init_message);
